Use tiny-fu instead of TinyScheme as the controlling program to run scripts.
The original plan to run a Scheme script was to have GIMP start TinyScheme
and TinyScheme would load tiny-fu as an extension. There were difficulties
in going from Scheme to C to Scheme (initially to create wrapper functions
for GIMP functions) so this approach has been abandoned in favour of the
original method of going from C to Scheme to C to run Scheme scripts. Also
updated dependencies to match those of GIMP. Don't use gtk+ 3.0 functions.